Refactor optimisation
Move project reduction into ProjectHelpers to apply it to package updates and tests too.
